refactor: rename collection::collection module to collection::core
Resolved clippy::module_inception warning.
