+++ rs/benchmarks/src/deletion_and_vacuum.rs
@@ -1,6 +1,6 @@
 use config::collection::CollectionConfig;
 use criterion::{BenchmarkId, Criterion, criterion_group, criterion_main};
-use index::collection::collection::Collection;
+use index::collection::core::Collection;
 use index::collection::reader::CollectionReader;
 use index::optimizers::vacuum::VacuumOptimizer;
 use quantization::noq::noq::NoQuantizerL2;
