+++ rs/benchmarks/src/insertion.rs
@@ -1,6 +1,6 @@
 use config::collection::CollectionConfig;
 use criterion::{BenchmarkId, Criterion, criterion_group, criterion_main};
-use index::collection::collection::Collection;
+use index::collection::core::Collection;
 use index::collection::reader::CollectionReader;
 use quantization::noq::noq::NoQuantizerL2;
 use tempdir::TempDir;
