+++ rs/index/src/optimizers/vacuum.rs
@@ -60,7 +60,7 @@ mod tests {
     use config::collection::CollectionConfig;
 
     use super::*;
-    use crate::collection::collection::Collection;
+    use crate::collection::core::Collection;
     use crate::collection::reader::CollectionReader;
     use crate::segment::Segment;
